I have fixed some of the styling issues with the footer, making it so that the sitemap section doesn't wrap, but I am currently unable to then find a way that the other text then does wrap.
I have fixed the issue with the 'Latest News' section. Where
the heights of the boxes kept constant which pushed the boxes into
each other. I have also fixed the margin of the title of that section.

I have been working on the DRY principle with my website, I had added
new variables for the 7 main colors used in the website, and tested
that my website still functions correctly after changing the code.
As well as tested that if you change 1 variable color,
it changes all of that color on the website.

Code has been seperated into partials.